Introduce FP16 (HalfFloat) tensor integration across TornadoVM layers and kernels.
- Adapt tensors, task graphs, and layer planners to support `HalfFloatArray`.
- Replace FP32 arrays with FP16-compatible implementations in key inference states (`wrapX`).
- Add new FP16-specific kernels for data transfer and activation computations.
- Optimize Q8_0 quantized operations with FP16 tensor support for improved efficiency.
- Update `State` classes and TornadoVM integrations to utilize FP16 data structures for key activation paths.

+++ src/main/java/org/beehive/gpullama3/inference/InferenceCore.java
@@ -583,7 +583,8 @@ public final class InferenceCore {
         final Configuration configuration = model.configuration();
         final TornadoWeights weights = (TornadoWeights) model.weights();
 
-        MemorySegment.copy(weights.getTokenEmbeddingTable().asFloatArray().getSegment(), (long) token * configuration.dim() * Float.BYTES, state.wrapX.getSegment(), 0, configuration.dim() * Float.BYTES);
+        //TODO: Xxxx
+        MemorySegment.copy(weights.getTokenEmbeddingTable().asFloatArray().getSegment(), (long) token * configuration.dim() * Short.BYTES, state.wrapX.getSegment(), 0, configuration.dim() * Short.BYTES);
 
         return tornadoVMMasterPlan.tornadoVMForwardExecuteLayered(position);
     }
